requirements.txt
pyTelegramBotAPI==4.19.1
Pillow==10.3.0
python-docx==1.1.2
flask==2.3.3
gunicorn==21.2.0